plugins/slate: alignment plugin
