+++ src/editor/plugins/content/slate/plugins/alignment.js
@@ -1,0 +1,39 @@
+/* eslint-disable prefer-reflect */
+import React from 'react'
+import AlignLeftIcon from 'material-ui/svg-icons/editor/format-align-left'
+import AlignCenterIcon from 'material-ui/svg-icons/editor/format-align-center'
+import AlignRightIcon from 'material-ui/svg-icons/editor/format-align-right'
+import AlignJustifyIcon from 'material-ui/svg-icons/editor/format-align-justify'
+
+import { ToolbarButton } from '../helpers'
+import Plugin from './Plugin'
+
+export default class HeadingsPlugin extends Plugin {
+  createButton = (align, icon) => ({ editorState, onChange }) => {
+    const onClick = (e) => {
+      e.preventDefault()
+
+      const isActive = editorState.blocks.some((block) => block.data.get('align') === align)
+
+      onChange(
+        editorState
+          .transform()
+          .setBlock({ data: { align: isActive ? null : align } })
+          .apply()
+      )
+    }
+
+    const isActive = editorState.blocks.some((block) => block.data.get('align') === align)
+
+    return <ToolbarButton onClick={onClick} isActive={isActive} icon={icon} />
+  }
+
+  name = 'alignment'
+
+  toolbarButtons = [
+    this.createButton('left', <AlignLeftIcon />),
+    this.createButton('center', <AlignCenterIcon />),
+    this.createButton('right', <AlignRightIcon />),
+    this.createButton('justify', <AlignJustifyIcon />)
+  ]
+}
